requirements.txt
flask
flask-cors
sqlalchemy
pandas
numpy
scikit-learn
xgboost
joblib
requests
python-dotenv
loguru
